+++ linkerd/examples/servicemesh.yaml
@@ -17,11 +17,7 @@ routers:
     - kind: io.l5d.port
       port: 4141
   dtab: |
-    /bb => /#/io.l5d.fs;
-    /aa/* => /$/inet/localhost/4140;
-    /b => /bb;
-    /a => /aa;
-    /svc => 2*/a & /b;
+    /svc => /#/io.l5d.fs;
   servers:
   - port: 4140
     ip: 0.0.0.0
